SDK-1544: Rename check report classes
