+++ src/DocScan/Request/Check/Report/SandboxBreakdown.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Yoti\Sandbox\DocScan\Request;
+namespace Yoti\Sandbox\DocScan\Request\Check\Report;
 
 use Yoti\Util\Json;
+use Yoti\Util\Validation;
 
 class SandboxBreakdown implements \JsonSerializable
 {
@@ -28,12 +29,14 @@ class SandboxBreakdown implements \JsonSerializable
      * SandboxBreakdown constructor.
      * @param string $subCheck
      * @param string $result
-     * @param SandboxDetails[] $details
+     * @param SandboxDetail[] $details
      */
     public function __construct(string $subCheck, string $result, array $details)
     {
         $this->subCheck = $subCheck;
         $this->result = $result;
+
+        Validation::isArrayOfType($details, [SandboxDetail::class], 'details');
         $this->details = $details;
     }
