+++ src/DocScan/Request/Check/Report/SandboxRecommendationBuilder.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yoti\Sandbox\DocScan\Request;
+namespace Yoti\Sandbox\DocScan\Request\Check\Report;
 
 class SandboxRecommendationBuilder
 {
